refactor: getContentType 메서드 이동
HttpResponse에 있던 getContentType 메서드를 더 관련이 있는 HttpRequest로
이동 시키고, 변경 사항을 RequestHandler와 테스트 클래스에 반영했습니다.
