+++ src/main/java/webserver/HttpResponse.java
@@ -10,17 +10,6 @@ public class HttpResponse {
 
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
 
-    public String getContentType(String path) {
-        logger.debug("Path : {}", path);
-        for (ContentType contentType : ContentType.values()) {
-            if (path.contains(contentType.getName())) {
-                return contentType.getProcess();
-            }
-        }
-        // contentType이 text/plain이거나 비어 있을 때, 해당 경로에 파일이 존재하지 않으면 정적 텍스트 파일로 간주하여, 다운로드를 진행
-        return "text/html";
-    }
-
     public void response200Header(DataOutputStream dos, int lengthOfBodyContent, String contentType) {
         try {
             dos.writeBytes("HTTP/1.1 200 OK \r\n");
